unmarshal

+++ packages/relayer/pkg/mock/eth_client.go
@@ -139,6 +139,6 @@ func (c *EthClient) SubscribeNewHead(ctx context.Context, ch chan<- *types.Heade
 }
 
 // nolint: lll
-func (c *EthClient) TransactionSender(ctx context.Context, tx *types.Transaction, blockHash common.Hash, txIndex uint8) (common.Address, error) {
+func (c *EthClient) TransactionSender(ctx context.Context, tx *types.Transaction, blockHash common.Hash, txIndex uint) (common.Address, error) {
 	return common.Address{}, nil
 }
